NET-49: fixed a bitset overflow bug

+++ consensus/packets/bitarray.go
@@ -86,7 +86,7 @@ func (arr *bitArray) serializeCompressed() ([]byte, error) {
 	if err != nil {
 		return nil, errors.Wrap(err, "[ serializeCompressed ] failed to get state from bitarray")
 	}
-	count := uint8(1)
+	count := uint16(1)
 	for i := 1; i < arr.bitsSize/2; i++ { // cuz 2 bits == 1 state
 		current, err := arr.getState(i)
 		if err != nil {
